Сделан блок парнеры, добавлена модульность
