+++ app/index.html
@@ -293,7 +293,7 @@
             </div>
         </section>
 
-        Our Insights & Articles
+        <!-- Our Insights & Articles -->
 
 
 
@@ -308,11 +308,51 @@
         </section>
     </main>
 
+    <section class="partners">
+        <div class="container">
+            <div class="partners__inner">
+                <ul class="partners__list">
+                    <li class="partners__list-item">
+                        <a class="partners__list-link" href="#">
+                            <img class="partners__list-img" src="images/partners/logo_partner-1.png"
+                                alt="logo Peterson & Stephanie">
+                        </a>
+                    </li>
 
+                    <li class="partners__list-item">
+                        <a class="partners__list-link" href="#">
+                            <img class="partners__list-img" src="images/partners/logo_partner-2.png"
+                                alt="logo KMN Consulting">
+                        </a>
+                    </li>
+
+                    <li class="partners__list-item">
+                        <a class="partners__list-link" href="#">
+                            <img class="partners__list-img" src="images/partners/logo_partner-3.png"
+                                alt="logo Chainforce">
+                        </a>
+                    </li>
+
+                    <li class="partners__list-item">
+                        <a class="partners__list-link" href="#">
+                            <img class="partners__list-img" src="images/partners/logo_partner-4.png" alt="logo NextG3N">
+                        </a>
+                    </li>
+
+                    <li class="partners__list-item">
+                        <a class="partners__list-link" href="#">
+                            <img class="partners__list-img" src="images/partners/logo_partner-5.png"
+                                alt="logo  The Vintage">
+                        </a>
+                    </li>
+                </ul>
+            </div>
+        </div>
+    </section>
 
 
     <footer class="footer">
-        <div class="container">
+        <div class="small-container">
             <div class="footer-top">
 
 
@@ -411,7 +451,7 @@
 
         </div>
     </footer>
-    <script src="js/main.min.js"></script>
+    <script src="js/main.min.js" type="module"></script>
 </body>
 
 </html>
