+++ app/js/view.js
@@ -1,0 +1,15 @@
+export const DOM = {
+    WEEK_ITEMS: document.querySelector('.week__items'),
+    ITEM_TEMPLATE: document.querySelector('#item-template'),
+}
+
+export const WEEK_PRODUCTS = [
+    { name: 'Wooden radio', price: '$29.00', image: 'images/products/week/product-1.png' },
+    { name: 'Wooden chair', price: '$29.00', image: 'images/products/week/product-2.png' },
+    { name: 'White chair', price: '$29.00', image: 'images/products/week/product-3.png' },
+    { name: 'Retro table', price: '$29.00', image: 'images/products/week/product-4.png' },
+    { name: 'Wooden radio', price: '$29.00', image: 'images/products/week/product-5.png' },
+    { name: 'Wooden chair', price: '$29.00', image: 'images/products/week/product-6.png' },
+    { name: 'White chair', price: '$29.00', image: 'images/products/week/product-7.png' },
+    { name: 'Retro table', price: '$29.00', image: 'images/products/week/product-8.png' },
+]
