+++ app/scss/_global.scss
@@ -21,3 +21,8 @@ body {
     margin: 0 auto;
     padding: 0 15px;
 }
+.small-container {
+    max-width: 1170px;
+    margin: 0 auto;
+    padding: 0 15px;
+}
