+++ app/scss/_vars.scss
@@ -6,7 +6,7 @@ $LIGHT-GREY: #edf1f4;
 $ADD-GREY: #7b7b7b;
 $ADD-GREY2:#b3b7b5;
 $LIGHT-BROWN: #fde7d3;
-$ADD-LIGHT-BROWN: #fef7f0;
+$ADD-LIGHT-BROWN: #fef7f0; 
 $ADD-LIGHT-BROWN2: #8b8b8b;
 $ADD-LIGHT-BROWN3: #feefe2;
 $BROWN: #fcd7b6;
